Fix setting visibility on hidden items
Until now there was a bug, where when there was a feature A whose parent
feature P was hidden, A's visibility could not be changed.

Now when A's visibility is changed while its parent is hidden, the
program will remember this so it is set correctly the moment P is set
visible again.