Allow mandatory features to be hidden
To allow more flexibility when declaring visibility constraints,
mandatory features can now be hidden again. But while they are hidden,
they still count as included in the model instance (this is important,
as other visibility constraints may depend on those features being
included).

This change is important as 1 mandatory feature might have many child
features, where only a subset needs to be selected. Often it makes sense
to hide the mandatory feature itself, and only show it, when the child
features become relevant.