Implement toggle to disable visibility constraints
A toggle button was added in the configurator page, to disable filtering
by visibility constraints. This is helpful, when the user wants to search
for a specific feature that might currently be hidden by visibility
constraints.

Now he can just click the button and all features will be displayed. If
the user wants to enable visibility constraint filtering again, he can
do so by pressing the button again.