Fix errors between autoselect and visibility
The autoselection of a feature can cause another feature to be set
visible if a corresponding visibility constraint exists.

Until now, this autoselection in combination with the setting of
visibility caused weird errors, where instead of being set visible it
was hidden instead, until the autoselection was reverted.

This behavior is now fixed.